Prepare for Redbooth tutorial.

+++ root/usr/local/lib/ml/utils.py
@@ -1,5 +1,5 @@
 import torch
 
 
-def tensor(X, requires_grad=False):
-    return torch.from_numpy(X).float().requires_grad_(requires_grad)
+def tensor(X, requires_grad=False, dtype=torch.float32):
+    return torch.from_numpy(X).to(dtype).requires_grad_(requires_grad)
